feat: close 함수 이동
